# docs/swagger.yaml
basePath: ./
definitions:
...
  dto.Error:
    properties:
      message:
        type: string
    type: object
  dto.IDWrapper:
    properties:
      id:
        minimum: 1
        type: integer
    required:
    - id
    type: object
...
  dto.PaginatedProductStockCollection:
    properties:
      data:
        items:
          $ref: '#/definitions/dto.ProductStock'
        type: array
      total:
        description: This will always return the total of all records
        type: integer
    type: object
  dto.PaginatedSupplierCollection:
    properties:
      data:
        items:
          $ref: '#/definitions/dto.Supplier'
        type: array
      total:
        description: This will always return the total of all records
        type: integer
    type: object
...
  dto.ProductStock:
    properties:
      id:
        type: integer
      product_id:
        type: integer
      stock_quantity:
        type: integer
    type: object
  dto.ProductStockUpdate:
    properties:
      product_id:
        type: integer
      stock_quantity:
        type: integer
    type: object
...
  dto.Supplier:
    properties:
      email:
        type: string
      id:
        type: integer
      is_verified_supplier:
        type: boolean
      name:
        type: string
      phone:
        type: string
      status_id:
        type: integer
    type: object
  dto.SupplierUpdate:
    properties:
      email:
        type: string
      id:
        type: integer
      is_verified_supplier:
        type: boolean
      name:
        type: string
      phone:
        type: string
      status_id:
        type: integer
    type: object
host: localhost:8080
info:
  contact: {}
  description: Describes technoStore REST API
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: technoStore API
  version: "1.0"
paths:
...
  /v1/product-stock:
    post:
      consumes:
      - application/json
      description: Create a new ProductStock in the system
      parameters:
      - description: ProductStock params
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.ProductStock'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.IDWrapper'
        "400":
          description: Invalid request body
          schema:
            type: string
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.Error'
        "500":
          description: Error
          schema:
            type: string
      summary: Add a new ProductStock
      tags:
      - ProductStock
  /v1/product-stock/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a ProductStock by id
      parameters:
      - description: ProductStock ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: ProductStock delete processed
          schema:
            type: string
        "400":
          description: Invalid request body
          schema:
            type: string
        "404":
          description: ProductStock not found
          schema:
            type: string
        "500":
          description: Error
          schema:
            type: string
      summary: Delete a ProductStock by id
      tags:
      - ProductStock
    get:
      consumes:
      - application/json
      description: Get a ProductStock by id
      parameters:
      - description: ProductStock ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.ProductStock'
        "400":
          description: Invalid request body
          schema:
            type: string
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.Error'
        "500":
          description: Error
          schema:
            type: string
      summary: Get a ProductStock by id
      tags:
      - ProductStock
    patch:
      consumes:
      - application/json
      description: Update a ProductStock by id
      parameters:
      - description: ProductStock params
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.ProductStockUpdate'
      - description: ProductStock ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: ProductStockDto updated
          schema:
            type: string
        "400":
          description: Invalid request body
          schema:
            type: string
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.Error'
        "500":
          description: Error
          schema:
            type: string
      summary: Update a ProductStock by id
      tags:
      - ProductStock
  /v1/product-stocks:
    get:
      consumes:
      - application/json
      description: Get ProductStocks
      parameters:
      - description: limit
        in: query
        name: limit
        type: integer
      - description: offset
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.PaginatedProductStockCollection'
        "400":
          description: Invalid request body
          schema:
            type: string
        "500":
          description: Error
          schema:
            type: string
      summary: Get ProductStocks
      tags:
      - ProductStock
...
  /v1/supplier:
    post:
      consumes:
      - application/json
      description: Create a new Supplier in the system
      parameters:
      - description: Supplier params
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.Supplier'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.IDWrapper'
        "400":
          description: Invalid request body
          schema:
            type: string
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.Error'
        "500":
          description: Error
          schema:
            type: string
      summary: Add a new Supplier
      tags:
      - Supplier
  /v1/supplier/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a Supplier by id
      parameters:
      - description: Supplier ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: Supplier delete processed
          schema:
            type: string
        "400":
          description: Invalid request body
          schema:
            type: string
        "404":
          description: Supplier not found
          schema:
            type: string
        "500":
          description: Error
          schema:
            type: string
      summary: Delete a Supplier by id
      tags:
      - Supplier
    get:
      consumes:
      - application/json
      description: Get a Supplier by id
      parameters:
      - description: Supplier ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.Supplier'
        "400":
          description: Invalid request body
          schema:
            type: string
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.Error'
        "500":
          description: Error
          schema:
            type: string
      summary: Get a Supplier by id
      tags:
      - Supplier
    patch:
      consumes:
      - application/json
      description: Update a Supplier by id
      parameters:
      - description: Supplier params
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.SupplierUpdate'
      - description: Supplier ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: SupplierDto updated
          schema:
            type: string
        "400":
          description: Invalid request body
          schema:
            type: string
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.Error'
        "500":
          description: Error
          schema:
            type: string
      summary: Update a Supplier by id
      tags:
      - Supplier
  /v1/suppliers:
    get:
      consumes:
      - application/json
      description: Get Suppliers
      parameters:
      - description: limit
        in: query
        name: limit
        type: integer
      - description: offset
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.PaginatedSupplierCollection'
        "400":
          description: Invalid request body
          schema:
            type: string
        "500":
          description: Error
          schema:
            type: string
      summary: Get Suppliers
      tags:
      - Supplier
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
  BasicAuth:
    type: basic
swagger: "2.0"